Move filter helper to add button, simplify templates file
