+++ client/src/components/Markdown/Editor/templates.ts
@@ -1,19 +1,8 @@
 import vegaBarChart from "./templates/vega-bar-chart.json";
 import vegaLineChart from "./templates/vega-line-chart.json";
-import type { CellType } from "./types";
+import type { TemplateCategory } from "./types";
 
-interface TemplateEntry {
-    title: string;
-    description: string;
-    cell: CellType;
-}
-
-interface TemplateCategory {
-    name: string;
-    templates: Array<TemplateEntry>;
-}
-
-export const templateCategories: Array<TemplateCategory> = [
+export const cellTemplates: Array<TemplateCategory> = [
     {
         name: "Markdown",
         templates: [
@@ -65,23 +54,3 @@ export const templateCategories: Array<TemplateCategory> = [
         ],
     },
 ];
-
-export function getTemplates(query: string): Array<TemplateCategory> {
-    const filteredCategories: Array<TemplateCategory> = [];
-    templateCategories.forEach((category) => {
-        const matchedTemplates = category.templates.filter(
-            (template) =>
-                category.name.toLowerCase().includes(query.toLowerCase()) ||
-                template.title.toLowerCase().includes(query.toLowerCase()) ||
-                template.description.toLowerCase().includes(query.toLowerCase())
-        );
-        if (matchedTemplates.length > 0) {
-            filteredCategories.push({
-                name: category.name,
-                templates: matchedTemplates,
-            });
-        }
-    });
-
-    return filteredCategories;
-}
