+++ client/src/components/Markdown/Editor/types.ts
@@ -3,3 +3,14 @@ export interface CellType {
     content: string;
     toggle?: boolean;
 }
+
+export interface TemplateCategory {
+    name: string;
+    templates: Array<TemplateEntry>;
+}
+
+export interface TemplateEntry {
+    title: string;
+    description: string;
+    cell: CellType;
+}
